Add test case for handling resend request loop
